// frontend/src/api/generated/model/scoutInvestorsResponseDtoDataItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */
import type { ScoutInvestorsResponseDtoDataItemApplicationStatus } from './scoutInvestorsResponseDtoDataItemApplicationStatus';

export type ScoutInvestorsResponseDtoDataItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  name: string;
  /** @pattern ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$ */
  email: string;
  /** @nullable */
  image?: string | null;
  hasApplied: boolean;
  /** @nullable */
  applicationStatus: ScoutInvestorsResponseDtoDataItemApplicationStatus;
};